+++ src/main/java/me/drex/vanish/util/VanishMixinConfigPlugin.java
@@ -1,5 +1,6 @@
 package me.drex.vanish.util;
 
+import net.fabricmc.loader.api.FabricLoader;
 import org.objectweb.asm.tree.ClassNode;
 import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
 import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
@@ -11,7 +12,6 @@ public class VanishMixinConfigPlugin implements IMixinConfigPlugin {
 
     @Override
     public void onLoad(String mixinPackage) {
-
     }
 
     @Override
@@ -21,6 +21,14 @@ public class VanishMixinConfigPlugin implements IMixinConfigPlugin {
 
     @Override
     public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
+        if (mixinClassName.startsWith("me.drex.vanish.mixin.compat.")) {
+            String compatName = mixinClassName.replace("me.drex.vanish.mixin.compat.", "");
+            String[] parts = compatName.split("\\.", 3);
+            String modId = parts[0];
+            String status = parts[1];
+            boolean enabled = status.equals("enabled");
+            return FabricLoader.getInstance().isModLoaded(modId) == enabled;
+        }
         return true;
     }